+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Allowing specific paths to be used instead of the search root.
 * Making the field picker work.
 * Made the facet component actually work.
+* Made the query component actually work.
 
 # 0.1.5